requirements.txt
selenium==4.15.0
beautifulsoup4==4.12.2
requests==2.31.0
Pillow==10.1.0
python-dotenv==1.0.0
lxml==4.9.3
aiohttp==3.9.0
playwright==1.40.0
pandas==2.1.3
pytest==7.4.3
webdriver-manager==4.0.1